# dircache: Add probabilistic validation for dramatic I/O reduction