--- a/etc/afpd/meson.build
+++ b/etc/afpd/meson.build
@@ -167,6 +167,7 @@ if have_dtrace and host_os != 'darwin'
                 'fork.c',
                 'appl.c',
                 'catsearch.c',
+                'dircache.c',
                 'directory.c',
                 'enumerate.c',
                 'file.c',