… doc/manpages/man5/afp.conf.5.md
… etc/afpd/dircache.c
… test/testsuite/lantest.c