Implement "ask forgiveness" model for directory cache validation that reduces filesystem stat() and lstat() calls by up to 99% while maintaining safety through automatic recovery and monitoring.

Relates to plan https://github.com/Netatalk/netatalk/discussions/2422. And resolves "Cache Validation Overhead" issue. This unblocks the work to generally modernise the dircache structure, and eventually implement Adaptive Replacement Cache (ARC).

User story to accompany this change; https://github.com/Netatalk/netatalk/wiki/DirCache-redundant-stat-ops

Key improvements:
- Configurable validation frequency (1-100) via afp.conf
- Smart detection distinguishes metadata from content changes
- Self-healing: Invalid entries automatically re-queried from CNID
- New metrics track effectiveness (expunged vs invalid_on_use)
- Backward compatible: Default freq=1 validates every access (By default nothing changes)

Performance impact:
- `dircache validation freq=10`: 90% reduction in stat() calls
- `dircache validation freq=100`: 99% reduction in stat() calls

Safety guarantees:
- Internal operations maintain immediate consistency
- External changes detected through periodic validation
- Invalid entries tracked and automatically repaired
- Observable metrics enable performance tuning

Significantly reduces storage I/O and page cache stress, improves AFP performance without compromising data integrity. Particularly beneficial for deployments where Netatalk is the sole accessor. 

Why does this does not make Netatalk 99x faster? The validate lstat/stat calls are usually served by the Linux Page Cache, which is very fast. So we are no longer doing something that was already very fast - as a result these changes will not be very noticable on modern hardware.. However, by not repeatedly calling lstat()/stat() (which hits the page cache) we avoid keeping unneeded entries warm in the page cache simply due to enumerating paths - this reduces page-cache LRU eviction, and avoids page-cache scanning - means the page cache can focus on more valuable and real hot objects.

---------------------------------------------------------------------------------

In real world testing I found an interesting issue - when renaming folders, and re-entering the renamed folder, most of the contents were missing (not shown as the cache entries were invalid). It turns out I think I stumbled on the original reason why the inital developer decided to just validate every single dircache entry every single access.. Doing so solves this rename invalidation problem..

In short, dircache entries include pointers to their parent entry path, and a chain of dircache entires are required to map the entire path from root to leaf (if parent entries do not exist, they are recursively built when first accessing a leaf). Therefore when you rename a folder in the middle of the path, the parent/path information in every single child dircache and CNID entry become invalid and stale.

This has always been true and the code to clean the cache/CNID trees after rename/move/delete did not exist (this is the root cause why the dircache and CNID database have historically always suffered with so many stale entries! - which can cause all sorts of strange issues - yay we found it..).

Therefore the second commit on this PR adds the missing logic to clean all child dircache entries during path rename/move/delete operations, and also fixed some old bugs where the code tries to delete dircache entries after file delete operations instead of before (root cause of many unexpected 'expunged' counters) :) So far it is working great, stale entries have been almost completely eradicated!

You can now rename a folder and access the renamed folder and everything is present and correct, the only difference is when accessing the renamed folder, it will be slower, as there is NO cache. The speed difference between the original folder and the renamed folder is the dircache :) This is especially noticable when renaming a folder halfway along a deep path.

This is much better as we now have a strictly synced dircache which is pruned and always up to date, to accompy this trust the cache change.

While working on this I discovered, as well as there being no code to prune the dircache tree, the same for the CNID is also missing. However the CNID API does not currently include an operation to search all entries by path to find items for pruning. This is a big change which I will omit for this PR, as CNID pruning has always been missing - and as long as the dircache entries are pruned, the resulting indirect CNID lookup to the stale record triggers a CNID update (so CNID is "eventually consistent" for path renames, but it realy should be strict synced).

So I will add the missing CNID API calls in another PR, as this same CNID API search operation we need need for Spotlight searching with CNID.

I am yet to experience any `invalid_on_use` counters, meaning the cache is always valid and correct (Netatalk only, no Samba, and no local file changes).


# Testing

## Alpine Linux Container

### No Changes (`dircache validation freq = 1`)
```
 /usr/local/bin/afp_lantest -h localhost -p 548 -u test -w test -s "File Sharing" -n 20
..
Netatalk Lantest Results (Averages and standard deviations (±) for all tests, across **20** iterations)
============================================================================================================

Test                                                                Time_ms  Time± AFPD_R AFPD_R± AFPD_W AFPD_W± CNID_R CNID_R± CNID_W CNID_W±   MB/s
------------------------------------------------------------------ -------- ------ ------ ------- ------ ------- ------ ------- ------ ------- ------
Open, stat and read 512 bytes from 1000 files [8,000 AFP ops]          3304  640.2   6000     0.0   7002     1.2      0     0.0      2     1.2      0
Writing one large file [103 AFP ops]                                     79   13.0      0     0.0    299     0.0      0     0.0      0     0.0   1265
Reading one large file [102 AFP ops]                                     37    2.8    100     0.0    100     0.0      0     0.0      0     0.0   2702
Locking/Unlocking 10000 times each [20,000 AFP ops]                     777    9.4      0     0.0  20000     0.0      0     0.0      0     0.0      0
Creating dir with 2000 files [4,000 AFP ops]                           3920  256.1   2000     0.0  10000     0.0      4     1.0   6130     7.3      0
Enumerate dir with 2000 files [~51 AFP ops]                            1129  247.6   1960     0.0     49     0.0      0     0.0      0     0.0      0
Deleting dir with 2000 files [2,000 AFP ops]                           3145  343.5   4000     0.0   4000     0.0      4     1.0   6179     4.6      0
Create directory tree with 1000 dirs [1,110 AFP ops]                   1734   65.8      0     0.0   4440     0.0      2     1.2   2273    20.5      0
Directory cache hits (100 dirs + 1000 files) [11,000 AFP ops]          3722 1265.8  10000     0.0  11100     0.0      0     0.0    100     0.0      0
Mixed cache operations (create/stat/enum/delete) [820 AFP ops]          626  110.1    820     0.0   1620     0.0      2     0.0   1221    24.2      0
Deep path traversal (nested directory navigation) [3,500 AFP ops]      1242  362.5   2500     0.0   3550     0.0      0     0.0     50     0.0      0
Cache validation efficiency (metadata changes) [30,000 AFP ops]        8508  856.1  30000     0.0  30100     0.0      0     0.0    100     0.0      0
------------------------------------------------------------------ -------- ------ ------ ------- ------ ------- ------ ------- ------ ------- ------
Sum of all AFP OPs = 80686                                        **28223**         57380          92260             12          16055

Aggregates Summary:
------------------------------------------------------------------
Average Time per AFP OP: 0.350 ms
Average AFPD Reads per AFP OP: 0.711
Average AFPD Writes per AFP OP: 1.143
See afp_lantest manpage for more information: https://netatalk.io/manual/en/afp_lantest.1

Dircache Statistics (/var/log/afpd.log):
------------------------------------------------------------------
Sep 30 02:23:14.893157 afpd[48] {dircache.c:797} (info:AFPDaemon): dircache statistics: (user: test) entries: 0, lookups: 2441264, 
**hits: 2280315 (93.4%)**, misses: 91841, **validations: ~2349423 (96.2%)**, added: 91909, removed: 91909, expunged: 69108, invalid_on_use: 0, evicted: 0, validation_freq: 1
```

### `dircache validation freq = 100`
```
Netatalk Lantest Results (Averages and standard deviations (±) for all tests, across **20** iterations)
============================================================================================================

Test                                                                Time_ms  Time± AFPD_R AFPD_R± AFPD_W AFPD_W± CNID_R CNID_R± CNID_W CNID_W±   MB/s
------------------------------------------------------------------ -------- ------ ------ ------- ------ ------- ------ ------- ------ ------- ------
Open, stat and read 512 bytes from 1000 files [8,000 AFP ops]          3095  573.5   6000     0.0   7000     0.3      0     0.0      1     0.0      0
Writing one large file [103 AFP ops]                                     85   18.7      0     0.0    299     0.0      0     0.0      0     0.0   1176
Reading one large file [102 AFP ops]                                     37    1.5    100     0.0    100     0.0      0     0.0      0     0.0   2702
Locking/Unlocking 10000 times each [20,000 AFP ops]                     767    9.9      0     0.0  20000     0.0      0     0.0      0     0.0      0
Creating dir with 2000 files [4,000 AFP ops]                           3498  194.7   2000     0.0  10000     0.0      4     1.0   6131     7.1      0
Enumerate dir with 2000 files [~51 AFP ops]                             950  212.1   1960     0.0     49     0.0      0     0.0      0     0.0      0
Deleting dir with 2000 files [2,000 AFP ops]                           2476  329.6   4000     0.0   4000     0.0      4     1.0   6179     4.7      0
Create directory tree with 1000 dirs [1,110 AFP ops]                   1536   45.3      0     0.0   4440     0.0      2     1.3   2277    26.7      0
Directory cache hits (100 dirs + 1000 files) [11,000 AFP ops]          2904  576.9  10000     0.0  11100     0.0      0     0.0    100     0.0      0
Mixed cache operations (create/stat/enum/delete) [820 AFP ops]          508   84.0    820     0.0   1620     0.0      2     0.0   1217    23.1      0
Deep path traversal (nested directory navigation) [3,500 AFP ops]       850  127.6   2500     0.0   3550     0.0      0     0.0     50     0.0      0
Cache validation efficiency (metadata changes) [30,000 AFP ops]        7804  271.2  30000     0.0  30100     0.0      0     0.0    100     0.0      0
------------------------------------------------------------------ -------- ------ ------ ------- ------ ------- ------ ------- ------ ------- ------
Sum of all AFP OPs = 80686                                        **24510**         57380          92258             12          16055               

Aggregates Summary:
------------------------------------------------------------------
Average Time per AFP OP: 0.304 ms
Average AFPD Reads per AFP OP: 0.711
Average AFPD Writes per AFP OP: 1.143
See afp_lantest manpage for more information: https://netatalk.io/manual/en/afp_lantest.1

Dircache Statistics (/var/log/afpd.log):
------------------------------------------------------------------
Sep 29 13:26:35.817496 afpd[36] {dircache.c:797} (info:AFPDaemon): dircache statistics: (user: test) entries: 0, lookups: 2441264, 
**hits: 2348728 (96.2%)**, misses: 91841, **validations: ~23494 (1.0%)**, added: 91844, removed: 91844, expunged: 695, invalid_on_use: 22572, evicted: 0, **validation_freq: 100**
```

The main take away is that we see a reduction in total runtime (28223ms down to 24510ms) and an increase in cache hits (93.4% to 96.2%).
Interestingly we see a much greater increase in the hit rate (delta = 68413) than we do for invalid_on_use (delta = 22572) - over 3x higher.

*This confirms, lower latency and higher hit rates with significantly less disk/page-cache IO*

## Lab Server

This has been running successfully on a lab server (FreeBSD) with various real-world client workloads, and no issues observed so far.

# Directory Cache Optimisation: "Trust the Cache"

## Probabilistic Cache Validation

This PR documents the new directory cache optimisation - Instead of validating every cache entry on every access (IO expensive), the Netatalk validates entries periodically and tracks when unvalidated entries turn out to be invalid during use.

### Key Changes

1. **Probabilistic Validation**: Cache entries validated every Nth access instead of every access
2. **Smart Validation**: Distinguishes metadata-only changes from content changes  
3. **Invalid Entry Tracking**: New `invalid_on_use` counter tracks unvalidated entries that fail
4. **Performance Optimisation**: Removed multiple redundant `lstat()` and `stat()` calls after cache lookup
5. **Configurable Parameters**: Admin-tunable validation frequency and metadata thresholds

---

## 1. Cache Entry Addition Flow

```mermaid
flowchart TD
    Start([File/Dir Access Request]) --> Check{Entry in<br/>Cache?}
    Check -->|No| Query[Query CNID Database]
    Check -->|Yes| Return1[Return Cached Entry]
    
    Query --> Create[Create struct dir]
    Create --> Store[Store inode & ctime]
    Store --> Add[dircache_add]
    Add --> Index1[Add to CNID Index]
    Index1 --> Index2[Add to DID/Name Index]
    Index2 --> Queue[Add to LRU Queue]
    Queue --> Return2[Return New Entry]
    
    style Start fill:#e1f5fe
    style Return1 fill:#c8e6c9
    style Return2 fill:#c8e6c9
```

### Key Points:
- Cache stores both **files** and **directories** (marked with `DIRF_ISFILE` flag)
- Each entry stores `dcache_ino` and `dcache_ctime` for validation
- Entries indexed three ways: CNID, DID/Name, and LRU queue

---

## 2. Probabilistic Validation Decision

```mermaid
flowchart TD
    Lookup([Cache Lookup]) --> Found{Entry<br/>Found?}
    Found -->|No| Miss[Cache Miss]
    Found -->|Yes| Validate{should_validate_cache_entry?}
    
    Validate --> Counter[Increment validation_counter]
    Counter --> Check{counter % freq == 0?}
    
    Check -->|Yes<br/>Validate Now| DoValidate[Perform Validation]
    Check -->|No<br/>Skip| SkipValidate[Skip Validation]
    
    DoValidate --> Stat[stat filesystem]
    Stat --> Valid{Entry<br/>Valid?}
    Valid -->|Yes| ClearFlag[Clear DIRF_UNVALIDATED<br/>Update ctime if needed]
    Valid -->|No| Remove1[dir_remove<br/>Mark as expunged]
    
    SkipValidate --> SetFlag[Set DIRF_UNVALIDATED flag]
    SetFlag --> Trust[Return Entry<br/>Trust Cache]
    
    ClearFlag --> ReturnValid[Return Valid Entry]
    Remove1 --> ReturnNull1[Return NULL]
    Miss --> ReturnNull2[Return NULL]
    
    style Trust fill:#fff3cd,stroke:#f0ad4e
    style ReturnValid fill:#c8e6c9
    style ReturnNull1 fill:#ffcdd2
    style ReturnNull2 fill:#ffcdd2
```

### Configuration Parameters:
- `dircache_validation_freq`: How often to validate (1-100)
- `dircache_metadata_window`: Time window for metadata-only changes (60-3600s)
- `dircache_metadata_threshold`: Max ctime delta for metadata-only (10-1800s)

---

## 3. Cache Entry Repair & Recovery

```mermaid
flowchart TD
    Use([Use Cached Entry]) --> Check{Entry<br/>Valid?}
    Check -->|Yes| Success[Operation Succeeds]
    
    Check -->|No| Failed[Operation Fails<br/>e.g., ENOENT]
    Failed --> Report{Has<br/>DIRF_UNVALIDATED?}
    
    Report -->|Yes| Count[Increment invalid_on_use]
    Report -->|No| Skip[Already counted as expunged]
    
    Count --> Remove[dir_remove]
    Skip --> Remove
    
    Remove --> Query[Query CNID Database]
    Query --> Found{Path<br/>Found?}
    
    Found -->|Yes| Rebuild[Rebuild Entry]
    Found -->|No| Error[Return Error]
    
    Rebuild --> NewStat[stat new path]
    NewStat --> Valid{Path<br/>Exists?}
    
    Valid -->|Yes| Create[Create New struct dir]
    Valid -->|No| Error
    
    Create --> AddCache[dircache_add]
    AddCache --> Return[Return New Entry]
    
    style Success fill:#c8e6c9
    style Count fill:#fff3cd,stroke:#f0ad4e
    style Return fill:#c8e6c9
    style Error fill:#ffcdd2
```

### Self-Healing Process:
1. Invalid entry detected during use
2. Entry removed from cache
3. CNID database queried for current path
4. New entry created and cached
5. Operation continues with fresh data

---

## 4. Invalid Entry Reporting Flow

```mermaid
flowchart TD
    Start([dir_remove Called]) --> Check{Entry Type}
    
    Check -->|ROOT| Skip[Skip Reporting]
    Check -->|Regular| Report[dircache_report_invalid_entry]
    
    Report --> Flag{Has DIRF_UNVALIDATED?}
    Flag -->|Yes| Inc[Increment invalid_on_use counter]
    Flag -->|No| NoInc[Don't increment<br/>Was validated]
    
    Inc --> Log[Log Debug Message]
    NoInc --> Continue
    Log --> Continue
    
    Continue[Continue Removal] --> Remove1[Remove from CNID index]
    Remove1 --> Remove2[Remove from DID/Name index]  
    Remove2 --> Remove3[Remove from LRU queue]
    Remove3 --> Queue[Add to invalid_dircache_entries]
    Queue --> Mark[Mark d_did = CNID_INVALID]
    Mark --> End([Entry Removed])
    
    style Inc fill:#fff3cd,stroke:#f0ad4e
    style NoInc fill:#d4edda
```

### Reporting Centralization:
- All invalid entries flow through `dir_remove()`
- Single point of responsibility for tracking
- Works for both files and directories

---

## 5. Model Comparison: Original vs "Ask Forgiveness"

### Original Model (Before Optimization)

```mermaid
stateDiagram-v2
    [*] --> CacheLookup: Request
    
    CacheLookup --> CacheHit: Found
    CacheLookup --> CacheMiss: Not Found
    
    CacheHit --> AlwaysValidate: ALWAYS stat()
    
    AlwaysValidate --> ValidEntry: Pass
    AlwaysValidate --> InvalidEntry: Fail
    
    InvalidEntry --> RemoveEntry: Remove
    
    RemoveEntry --> Recover: Query CNID
    
    Recover --> CacheRebuild: Path Found
    Recover --> Error: Path Not Found
    
    CacheRebuild --> [*]: Success
    ValidEntry --> [*]: Success
    CacheMiss --> Recover: Query CNID
    Error --> [*]: Failure
    
    note right of AlwaysValidate: Every cache hit<br/>causes filesystem I/O
```

### New "Ask Forgiveness" Model (After Optimization)

```mermaid
stateDiagram-v2
    [*] --> CacheLookup: Request
    
    CacheLookup --> CacheHit: Found
    CacheLookup --> CacheMiss: Not Found
    
    CacheHit --> ProbabilisticCheck: Check Frequency
    
    ProbabilisticCheck --> Validated: Validate Now (1/N)
    ProbabilisticCheck --> Unvalidated: Skip (N-1/N)
    
    Validated --> ValidEntry: Pass
    Validated --> InvalidEntry: Fail
    
    Unvalidated --> TrustAndUse: Mark DIRF_UNVALIDATED
    
    TrustAndUse --> UseSucceeds: Valid
    TrustAndUse --> UseFails: Invalid
    
    UseFails --> ReportInvalid: Count as invalid_on_use
    InvalidEntry --> RemoveEntry: Count as expunged
    
    ReportInvalid --> RemoveEntry
    RemoveEntry --> Recover: Query CNID
    
    Recover --> CacheRebuild: Path Found
    Recover --> Error: Path Not Found
    
    CacheRebuild --> [*]: Success
    ValidEntry --> [*]: Success
    UseSucceeds --> [*]: Success
    CacheMiss --> Recover: Query CNID
    Error --> [*]: Failure
    
    note right of Unvalidated: Most cache hits<br/>avoid filesystem I/O
```

### Key Differences:
…